fix(uffd): divide REMOVE byte range by pageSize before uint32 cast
uint32(rm.end-rm.start) truncated the byte length to 32 bits before
dividing by pageSize, silently dropping pages from REMOVE events larger
than 4 GiB. Divide first; uint32 block indices still address up to
~16 TiB at 4 KiB pages.

+++ packages/orchestrator/pkg/sandbox/uffd/userfaultfd/userfaultfd.go
@@ -289,7 +289,7 @@ func (u *Userfaultfd) Serve(
 					}
 
 					startIdx := uint32(header.BlockIdx(startOff, int64(u.pageSize)))
-					endIdx := startIdx + uint32(rm.end-rm.start)/uint32(u.pageSize)
+					endIdx := startIdx + uint32(uint64(rm.end-rm.start)/uint64(u.pageSize))
 					u.pageTracker.SetRange(startIdx, endIdx, block.Zero)
 				}
 				u.settleRequests.Unlock()
